[process/nginx] fixes #43 by ordering hashes
Puppet 2.7 that we support runs on ruby 1.8.5 or 1.8.7
:crying_cat_face:
As a consequence, the hash when serialized by the .to_yaml function
can go in any order in the rendered string.
To avoid that we monkey patch the to_yaml function to sort the hash
by key every time if the ruby we are using is < 1.9.
It should avoid puppet to regenerate the same resource over and over
just because of some stupid ordering issues.