Allow multiple instances in the http_check.
Adds the ``instances`` parameter to the ``http_check`` class.

This parameter defaults to undef.

For backwards compatibility, if the ``instances`` array not defined it
will be instantiated as an array containing a single hash. The remaining
parameters of the class will be used for the key/value pairs in the
hash. This will allow us to iterate over the ``instances`` array in the
template.