Raise the priority for 'changed' event types so they're visible in the default UI
If your Puppet infrastructure never fails, let's say because you have a good
testing pipeline, it might not be obvious to you that any Puppet events are
being reported because everything but failures is 'low' priority. Datadog's web
UI defaults to showing only things of 'normal' priority.

This changes the priority of a 'changed' resource to 'normal' since a change of
configuration is useful information to use when correlating data or showing
events in a dashboard.