Validate the process hashes for the process integration
This adds a custom function that validates all fields of the
`$processes` array of hashes from the process integration. It had to be
done in a custom function in order to support all specific versions of
Puppet, otherwise it could be done with iterators at the manifest level.

Feel free to reject this if it's not the sort of thing you want to
maintain in this repo, this just bit me earlier today, so I thought I
would open this and see :)