Support ports as integers
Without this patch applied, compilation fails if integer data for ports
(proxy_port, pup_port etc) is passed in.

This is a problem, because it requires us to quote ports in Hiera, and
is a surprising and unexpected behaviour.

After this patch is applied, it is possible to specify these ports
either as a string or an integer.

Tests are added for each port showing that integer or string inputs both
lead to the same configuration in the catalog.