Hacky Bug Fix to support Amazon Linux and Datadog agent 6
This isn't a good fix. The problem is that Amazon Linux doesn't have systemd,
but it does have Upstart. The Datadog RPM installs an upstart configuration, but
Puppet tries to use the service command to start/stop/status the service which
fails on Amazon Linux.

By setting the provider to Upstart, we can ensure that on Amazon Linux Puppet
is able to start the agent using Upstart correctly.

Personally I'd consider this a bug in Amazon Linux really, but it's going to
fix itself when people move to Amazon Linux 2 that was recently released which
now moves to using systemd.