fix scoping bug
previously, the ca_certs conditional was nested in the 'tls cert
check is disabled' conditional.  while it doesn't make a _ton_ of sense
to provide a ca cert when you aren't checking it, allowing users to do
so makes migrating from an unchecked to a checked workflow easier.  I
don't see a downside to allowing this behavior.