Remove identation of the pgbouncer template to match the needed format of configuration file
With the current template the output of the configuration if the tags field was set was wrongly indented:

Output Example:

```
init_config:

instances:
  - host: localhost
    port: 5678
    username: datadog
    password: some-interesting-password
      tags:
      - dbinstanceidentifier:some-instance-identifier
```

The agent in this scenario was working right but if the integration with pgbouncer was not working as the tag field was wrongly indented.

With this change, the format will be the following

```
init_config:

instances:
  - host: localhost
    port: 5678
    username: datadog
    password: some-interesting-password
    tags:
      - dbinstanceidentifier:some-instance-identifier
```

Making the integration work if tags are specified.