Give Administrator group access explicitly
On Windows, Administrator doesn't have access to every file, and if we
remove access to the Administrator user or group, we lock ourself out of
the file. In other places this is already taken into account and explicitly
set, but here we were using the default value for mode. The default worked
well in previous tests, but I run into the lock out problem on a test using
Windows 2008 with puppet 6.10, so better make this explicit here too.