Migrate to PDK (#597)
Puppet Development Kit (PDK) is the recommended way to build and release Puppet modules.

PDK adds a base set of files to the projects (think skel in the Unix world), which I've tried to respect and merge with the stuff we already had.

Gem dependencies are now grouped under the meta-packages `puppet-module-*`, although unfortunately this is not available for Ruby 2.2 and earlier, where we still have to list dependencies manually. Also, these meta-packages don't work with unsupported combinations like Ruby 2.6 with Puppet < 6, so I had to remove those from the test matrix.

I haven't enabled the rubocop check since it has lots of offenses, but will do so on a separate PR.

Releases of the module can now be done using `pdk build`.