Allow digging into fields of facts when they are a hash
If a `facts_to_tags` tag looks like a hash path (eg: 'networking.domain'),
dig that path into the hash and use the inner value for the tag.

For backwards compatibility, if the name looks like a hash path
but there's a fact that's an exact string match (eg: there's a fact named
'networking.domain') then that will be used for the tag.