Only declare the Agent package if not already declared (#672)
We need the Agent package to be declared even when we don't install it,
since there's code that depends on it. Our current approach is to always
declare it, which works if the package is installed by something other than
Puppet (eg: chocolatey), since our declaration would not try to install it
again and the code that depends on that declaration would still be happy.
However, in case the package is being installed by a different Puppet
module, it would fail with a Duplicate declaration error.